// src/types/rider.ts
// ─── Rider ───────────────────────────────────────────────────────────────────
export interface Rider {
    rider_id: string;
    rider_name: string;
    rider_contact_num?: string;
}